HashMap Generated Chunks
Changed generatedChunks in WorldGeneration.java from an ArrayList to a
HashMap.  This prevents unnecessary looping when checking to see if a
chunk exists or not.  This will also be useful in checking adjacent
blocks when they exist on the edge of a chunk.

Note:  Commented out tree creation for the time being since it currently isn't set up to work with the current chunk system.