Performance Optimization
This commit increases the performance of the game multiple times over.
One of the main performance boosts was caching the texture(s) instead of
running Loader.loadTexture over and over again. Also made changes with
how blocks/chunks are created and rendered.  Fixed the problems that
existed when checking for adjacent blocks.  Fixed the camera W/S
movement.  Every single block in the Blocks package is currently in a
non-finished or barely-working state.  Those are next on the fix-list.

TO-DO:  Create a spriteSheet.PNG that has every texture we will need for our game.  Then, using SpriteSheetLoader.java, allow the blocks to run a function that will return the correct UV coordinates and more.