Chunk Animation/Player Movement and Collision/UI Tweaks Update
Added a scaling animation for chunks as they load in.  Also added
spacebar, A, D, and ctrl inputs for movement.  Cleaned up a few things
in MainGame and limited the size of the possible seed input to prevent
crashes.

Note:  Certain things in WorldGeneration are very broken.  Example:  Parts of trees despawn when you move from the the spawning chunk.  This will hopefully be fixed soon.