PlayerHand, Font, and UI Changes
Finished the texturing, scaling, and positioning of the player hand.
Added logic so player can't place block if they don't have a block in
their hand.  I also re-positioned some UI elements, added a button click
sound to the play button, and added a minecraft font to the title page.